requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
segno==1.5.2
python-slugify==8.0.1
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
python-multipart==0.0.6
